+++ packages/deploy/src/checks.ts
@@ -2,7 +2,7 @@ import { ProxyAdmin__factory, Wallet__factory } from "@nocturne-xyz/contracts";
 import { ethers } from "ethers";
 import { NocturneDeploymentConfig } from "./deployment";
 import { proxyAdmin, proxyImplementation } from "./proxyUtils";
-import { assert } from "./utils";
+import { assertOrErr } from "./utils";
 
 export async function checkNocturneDeploymentConfig(
   deployment: NocturneDeploymentConfig,
@@ -14,7 +14,7 @@ export async function checkNocturneDeploymentConfig(
     provider
   );
   const proxyAdminOwnerActual = await proxyAdminContract.owner();
-  assert(
+  assertOrErr(
     proxyAdminOwnerActual === deployment.proxyAdminOwner,
     "Proxy admin owner inconsistent"
   );
@@ -24,7 +24,7 @@ export async function checkNocturneDeploymentConfig(
     provider,
     deployment.walletProxy.proxy
   );
-  assert(
+  assertOrErr(
     walletProxyAdmin === deployment.proxyAdmin,
     "Wallet proxy admin incorrectly set"
   );
@@ -34,7 +34,7 @@ export async function checkNocturneDeploymentConfig(
     provider,
     deployment.vaultProxy.proxy
   );
-  assert(
+  assertOrErr(
     vaultProxyAdmin === deployment.proxyAdmin,
     "Vault proxy admin incorrectly set"
   );
@@ -44,7 +44,7 @@ export async function checkNocturneDeploymentConfig(
     provider,
     deployment.walletProxy.proxy
   );
-  assert(
+  assertOrErr(
     walletProxyImplementation === deployment.walletProxy.implementation,
     "Wallet proxy implementation does not match deployment"
   );
@@ -54,7 +54,7 @@ export async function checkNocturneDeploymentConfig(
     provider,
     deployment.vaultProxy.proxy
   );
-  assert(
+  assertOrErr(
     vaultProxyImplementation === deployment.vaultProxy.implementation,
     "Vault proxy implementation does not match deployment"
   );
@@ -66,7 +66,7 @@ export async function checkNocturneDeploymentConfig(
 
   // Wallet joinsplit verifier matches deployment
   const walletJoinSplitVerifier = await walletContract._joinSplitVerifier();
-  assert(
+  assertOrErr(
     walletJoinSplitVerifier === deployment.joinSplitVerifier,
     "Wallet joinsplit verifier does not match deployment"
   );